djangoapp/requirements.txt

Django>=5.1.3,<5.2
pillow>=11.0.0,<11.1
pytz>=2024.2,<2024.3
sqlparse>=0.5.2,<0.6
tzdata>=2024.2,<2024.3
psycopg2-binary>=2.9.10,<2.10
django-summernote>=0.8.20.0,<0.8.21
python-dotenv>=1.0.0,<1.1
django-axes>=6.0.1,<6.1



